requirements.txt
tensorflow==1.7.*
Pillow>=4.2.1
matplotlib
numpy>=1.11.0
jupyter
pytest>=3.2.2
docopt
pyyaml
protobuf==3.5.2
grpcio==1.11.0
pytorch==1.*
pandas
scipy
ipykernel
Python==3.6.*
unityagents==0.4.0
commentjson